+++ lib/rom/transaction.rb
@@ -1,0 +1,24 @@
+module ROM
+  # @api private
+  class Transaction
+    # @api private
+    Rollback = Class.new(StandardError)
+
+    # @api private
+    def run(opts = EMPTY_HASH)
+      yield(self)
+    rescue Rollback
+      # noop
+    end
+
+    # Unconditionally roll back the current transaction
+    #
+    # @api public
+    def rollback!
+      raise Rollback
+    end
+
+    # @api private
+    NoOp = Transaction.new.freeze
+  end
+end